requirements.txt
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
cryptography==35.0.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
mysql-connector-python==8.0.27
numpy==1.21.3
outcome==1.1.0
pandas==1.3.4
protobuf==3.19.1
pycparser==2.20
PyMySQL==1.0.2
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.29
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
